+++ src/App.js
@@ -13,11 +13,11 @@ import { Landing } from './pages/Landing.js';
 import { ComingSoon } from './pages/ComingSoon.js';
 // import Catalog from './pages/catalog.js';
 import Catalog from './pages/Catalog.js';
-import { SignUp } from './pages/SignUp.js';
-import { SignIn } from './pages/SignIn.js';
-import RegisterStudent from './pages/RegisterStudent.js';
-import { ResetPassword } from './pages/ResetPassword.js';
-import { ResetPassword2 } from './pages/ResetPassword2.js';
+import { SignUp } from './pages/clubLogin/ClubRegister.js';
+import { SignIn } from './pages/clubLogin/ClubLogin.js';
+import RegisterStudent from './pages/student/RegisterStudent.js';
+import { ResetPassword } from './pages/clubLogin/ResetPassword.js';
+import { ResetPassword2 } from './pages/clubLogin/ResetPassword2.js';
 import Security from './pages/admin/security/Security.js';
 import ErrorPage from './pages/ErrorPage';
 import AboutPage from './pages/AboutPage';
@@ -25,7 +25,7 @@ import About from './pages/AboutPage';
 import FAQ from './pages/FAQ';
 import Admin from './pages/admin/admin/Admin.js';
 import ClubPage from './pages/ClubPage';
-import Dashboard from './pages/Dashboard.js';
+import Dashboard from './pages/student/Dashboard.js';
 import Favorites from './pages/Favorites.js';
 import Activation from './pages/Activation.js';
 import store from './redux/store';
