+++ src/components/layout/about/PolaroidComponent.css
