+++ src/components/layout/richText/RichText.js
