+++ src/pages/Activation.js
@@ -4,7 +4,7 @@ import { withRouter } from 'react-router-dom';
 import signup from './assets/signup.png';
 import Dropdown from './Dropdown.js';
 import { updateProfile } from '../redux/actions/profile';
-import './SignUp.css';
+import './clubLogin/ClubRegister.css';
 import './Activation.css';
 import Loading from '../components/layout/loading/Loading';
 import { NotificationManager } from 'react-notifications';
