+++ src/pages/student/studentOnboarding/OnboardingFavorites.css
